Pass more events from the fragment presenters to the service presenter.

+++ app/src/main/java/com/colinchartier/overture/app/services/impl/DefaultMusicPresenter.java
@@ -42,7 +42,7 @@ public class DefaultMusicPresenter implements MusicPresenter {
     }
 
     @Override
-    public void seekToPercent(int percent) {
+    public void seekToPercent(double percent) {
 
     }
